refactor(subscription): move mock into subscription-api + isMockProvider on interface
Address review feedback:
- Add isMockProvider to the SubscriptionProvider interface (default false;
  MockSubscriptionProvider overrides true), so SubscriptionRepository no
  longer type-checks the concrete class.
- Move MockSubscriptionProvider into the subscription-api contract module
  next to NoOpSubscriptionProvider. It stays dependency-free: persistence
  (read/write lambdas) and the app ids (premium access, credit-pack prefix
  and placement) are injected by shared DI, which wires UserPreferences +
  Constants. The mock-detection helper (isSubscriptionMockActive, reads
  BuildConfig) stays in shared.
- Move the demo banner to the design system as a generic DemoBanner(text)
  component (resource-free); the paywall screens pass the localized string.
