+++ MobileApp/designsystem/src/commonMain/kotlin/com/kotlinfoundation/koko/designsystem/components/DemoBanner.kt
@@ -1,4 +1,4 @@
-package com.kotlinfoundation.koko.presentation.screens.paywall.components
+package com.kotlinfoundation.koko.designsystem.components
 
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.shape.RoundedCornerShape
@@ -6,30 +6,37 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import com.kotlinfoundation.koko.designsystem.components.AppCardContainer
 import com.kotlinfoundation.koko.designsystem.theme.AppTheme
-import com.kotlinfoundation.koko.generated.resources.Res
-import com.kotlinfoundation.koko.generated.resources.paywall_demo_banner
-import org.jetbrains.compose.resources.stringResource
+import com.kotlinfoundation.koko.designsystem.util.PreviewHelper
 
 /**
- * A loud, always-on notice shown at the top of the paywall while the mock subscription provider is
- * active (no real key set). Makes it impossible to mistake simulated purchases for real ones.
+ * A loud, warning-tinted notice strip. Used to mark a screen as running in a demo/simulated mode
+ * (e.g. the paywall while the mock subscription provider is active) so it can't be mistaken for real.
+ * The [text] is passed in so this stays a generic, resource-free design-system component.
  */
 @Composable
-internal fun DemoPaywallBanner(modifier: Modifier = Modifier) {
+fun DemoBanner(text: String, modifier: Modifier = Modifier) {
     AppCardContainer(
         modifier = modifier.fillMaxWidth(),
         shape = RoundedCornerShape(12.dp),
         backgroundColor = AppTheme.colors.status.warningContainer,
     ) {
         Text(
-            text = stringResource(Res.string.paywall_demo_banner),
+            text = text,
             style = AppTheme.typography.bodySmall,
             color = AppTheme.colors.text.primary,
             textAlign = TextAlign.Center,
             modifier = Modifier.fillMaxWidth(),
         )
+    }
+}
+
+@Composable
+@Preview
+internal fun DemoBannerPreview() {
+    PreviewHelper {
+        DemoBanner(text = "Demo mode — purchases are simulated.")
     }
 }
