+++ MobileApp/shared/src/commonMain/kotlin/com/kotlinfoundation/koko/data/repository/SubscriptionRepository.kt
@@ -2,7 +2,6 @@ package com.kotlinfoundation.koko.data.repository
 
 import com.kotlinfoundation.koko.data.BackgroundExecutor
 import com.kotlinfoundation.koko.data.source.preferences.UserPreferences
-import com.kotlinfoundation.koko.data.subscription.MockSubscriptionProvider
 import com.kotlinfoundation.koko.domain.model.Subscription
 import com.kotlinfoundation.koko.subscription.api.BillingPeriod
 import com.kotlinfoundation.koko.subscription.api.PurchasePackage
@@ -80,9 +79,9 @@ class SubscriptionRepository(
 
     /**
      * True when the active provider is the built-in mock (no real subscription key configured).
-     * The paywall shows a "Demo" banner in this case. See [MockSubscriptionProvider].
+     * The paywall shows a "Demo" banner in this case.
      */
-    val isMockProvider: Boolean get() = subscriptionProvider is MockSubscriptionProvider
+    val isMockProvider: Boolean get() = subscriptionProvider.isMockProvider
 
     suspend fun hasPremiumAccess(): Boolean = hasAccess(PAYWALL_PREMIUM_ACCESS)
